fix: set encoding to UTF-8 and disable newline conversion
Haskell supports automatically doing text encoding when writing
to a Handle. Unfortunately, it determines which encoding to use
by default by checking the system locale. This is a classic source
of headaches.

Notably, the ubuntu-latest seems to set its locale to en_US, which
uses ISO-8859-1, and *not* en_US.UTF-8. We could set this as part
of the action, but the more robust thing is to always use UTF-8.

I've also taken the pre-emptive step of disabling newline conversion
as well: this shouldn't be an issue for us currently, but will become
relevant if/when we support windows.