fix: use lowercase type variable names
This works around a grammatical oddity of Idris 2. Idris will happily
accept the following definition

```idris2
f : {A : Type} -> {x0, x1, x2, x3, x4, x5 : A} -> (y : A) -> A
f {A = a} y = y
```

However, it will *not* accept the following:

```idris2
f : {A : Type} -> {x0, x1, x2, x3, x4, x5 : A} -> (y : A) -> A
f {A = A} y = y
```

This is because Idris 2 thinks that the A on the RHS of the equality
ought to be a data constructor.